Crate consensus_config

Structs

Authority

Represents one authority in the committee.

AuthorityIndex

Each authority is uniquely identified by its AuthorityIndex in the Committee. AuthorityIndex is between 0 (inclusive) and the total number of authorities (exclusive).

Committee

Committee is the set of authorities that participate in the consensus protocol for this epoch. Its configuration is stored and computed on chain.

Parameters

Operational configurations of a consensus authority.

Constants

DIGEST_LENGTH

INTENT_MESSAGE_LENGTH

Type Definitions

DefaultHashFunction

For block digest.

Epoch

Committee of the consensus protocol is updated each epoch.

NetworkKeyPair

NetworkPrivateKey

NetworkPublicKey

Network key signs network messages and blocks.

ProtocolKeyPair

ProtocolPrivateKey

ProtocolPublicKey

Protocol key is used in random beacon.

ProtocolPublicKeyBytes

Stake

Voting power of an authority, roughly proportional to the actual amount of Sui staked by the authority. Total stake / voting power of all authorities should sum to 10,000.