Crate consensus_config

Structs

AnemoParameters

Authority

Represents one authority in the committee.

AuthorityIndex

Each authority is uniquely identified by its AuthorityIndex in the Committee. AuthorityIndex is between 0 (inclusive) and the total number of authorities (exclusive).

AuthorityKeyPair

AuthorityPublicKey

Authority key represents the identity of an authority. It is only used for identity sanity checks and not used for verification.

Committee

Committee is the set of authorities that participate in the consensus protocol for this epoch. Its configuration is stored and computed on chain.

NetworkKeyPair

NetworkPrivateKey

NetworkPublicKey

Network key is used for TLS and as the network identity of the authority.

Parameters

Operational configurations of a consensus authority.

ProtocolKeyPair

ProtocolKeySignature

ProtocolPublicKey

Protocol key is used for signing blocks and verifying block signatures.

TonicParameters

Constants

DIGEST_LENGTH

INTENT_MESSAGE_LENGTH

Functions

local_committee_and_keys

Creates a committee for local testing, and the corresponding key pairs for the authorities.

Type Aliases

DefaultHashFunction

Defines algorithm and format of block and commit digests.

Epoch

Committee of the consensus protocol is updated each epoch.

Stake

Voting power of an authority, roughly proportional to the actual amount of Sui staked by the authority. Total stake / voting power of all authorities should sum to 10,000.